major refactoring
This contains a major revision to the driver code. The changes have been made partly to reflect the new transaction mechanism, which leverages the LiUGraphQL/graphql-js lib, to clean up sections of the driver that contained reduncencies, and to make all constraint checks happen as part of the ongoing transaction, and to make the code easier to follow.